renamed regions jsons to be more specific what codesets they include
